% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa_info_db}
\alias{aa_info_db}
\title{Example amino acid information db}
\format{A data.frame with 519 protein IDs and the connected genbank IDs. 
\describe{
  \item{1st}{Protein ID}
  \item{2nd}{Genbank ID}
  \item{3rd}{Protein start position}
  \item{4th}{Protein end position}
  \item{5th}{Identifier}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/genbank}
}
\usage{
aa_info_db
}
\description{
A dataset containing 519 protein IDs, the start and the end position
on the DNA sequence
}
\keyword{datasets}
