% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_subset_index.R
\name{build_index}
\alias{build_index}
\title{Test}
\usage{
build_index(dna_set, aa_set, index_dir = NULL, index_name = NULL,
  bowtie = TRUE, pauda = TRUE, star = FALSE, force = FALSE,
  tmp_dir = tempdir())
}
\arguments{
\item{dna_set}{}

\item{aa_set}{}

\item{index_dir}{}

\item{index_name}{}

\item{bowtie}{}

\item{pauda}{}

\item{star}{}

\item{force}{}

\item{tmp_dir}{}
}
\value{
file.list
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
