% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{build_sample_info}
\alias{build_sample_info}
\title{Test}
\usage{
build_sample_info(sample_in, out_file = NULL, par_list, proc_start_tm)
}
\arguments{
\item{sample_in}{}

\item{out_file}{}

\item{par_list}{}

\item{proc_start_tm}{}
}
\value{
data.frame
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
