% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_host}
\alias{check_host}
\title{Experimental function to check the host by blastn}
\usage{
check_host(file, force = FALSE)
}
\arguments{
\item{file}{Infile of the sample, see \code{\link{get_sub_files}}.}

\item{force}{Should all files be overwritten?}
}
\value{
data.frame
}
\description{
A experimental function, which checks the host using blastn
}
\details{
This very experimental function uses blast and a subset of 100
read to determine the host of the sample. This works sometimes,
but it is very time consuming, with sometimes little gain.
}
\author{
Jochen Kruppa
}
