% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{collapse_df}
\alias{collapse_df}
\title{Collapse a matrix by row}
\usage{
collapse_df(mat, sep = "_")
}
\arguments{
\item{mat}{A matrix}

\item{sep}{A seperator for collapsing}
}
\value{
matrix
}
\description{
Helper function for collapsing a matrix
}
\details{
Helper function for collapsing a matrix
}
\author{
Jochen Kruppa
}
