% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dna_seqs}
\alias{dna_seqs}
\title{DNAStringSet of DNA sequences}
\format{A named DNAStringSet with 287 viral entries. The names are split by ' ' meaning:
\describe{
  \item{1st}{Genbank ID}
  \item{2nd}{Genbank ID}
  \item{3rd - End}{Description}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/genbank}
}
\usage{
dna_seqs
}
\description{
A DNAStringSet containing 287 DNA sequences from the NCBI GenBank.
}
\keyword{datasets}
