% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_obj.R
\name{empty_map_pep_list}
\alias{empty_map_pep_list}
\title{Internal empty object for pepe mapping}
\usage{
empty_map_pep_list(...)
}
\arguments{
\item{...}{No parameters needed}
}
\value{
data.frame
}
\description{
Internal empty object for pepe mapping
}
\details{
Internal empty object for pepe mapping
}
\author{
Jochen Kruppa
}
