% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{fastq_trimmer}
\alias{fastq_trimmer}
\title{Test}
\usage{
fastq_trimmer(inFile, outFile, leading = 10, trailing = 10,
  minlength = 50, illuminaclip, log_file)
}
\arguments{
\item{inFile}{}

\item{outFile}{}

\item{leading}{}

\item{trailing}{}

\item{minlength}{}

\item{illuminaclip}{}

\item{log_file}{}
}
\value{
Test
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
