% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fq2fa}
\alias{fq2fa}
\title{A wrapper for seqtk to convert fastq to fasta}
\usage{
fq2fa(fqFile, faFile)
}
\arguments{
\item{fqFile}{File path to fastq infile}

\item{faFile}{File path to fasta outfile}
}
\description{
A wrapper for seqtk to convert fastq to fasta
}
\details{
The external program seqtk must be installed and the path must be
set in the program_list().
}
\author{
Jochen Kruppa
}
