% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bowtie2_cmd.R
\name{get_bowtie2_cmd}
\alias{get_bowtie2_cmd}
\title{Wrapper to build up the Bowtie2 command}
\usage{
get_bowtie2_cmd(inFile, samOutFile, referenceDir, method, p = 32,
  all = FALSE)
}
\arguments{
\item{inFile}{File path to the in file(s)}

\item{samOutFile}{Path to the sam output file}

\item{referenceDir}{Path to the reference dir}

\item{method}{Should ene-to-end or lcal mapping be used?}

\item{p}{Number of used cores}

\item{all}{Shoul the reads mapped to all possible positions on the reference [default = FALSE]}
}
\value{
character
}
\description{
Wrapper function to build up the Bowtie2 command for the the system() call
}
\details{
This function is running internally in the \code{\link{map_dna_ref}}.
}
\author{
Jochen Kruppa
}
