% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{get_consensus_df}
\alias{get_consensus_df}
\title{Get the consenus sequence of the read hits to the reference}
\usage{
get_consensus_df(hits, out_file, par_list)
}
\arguments{
\item{hits}{Genebank IDs of the hits}

\item{out_file}{File path to the results dir}

\item{par_list}{Parameter given by the par_list(), see
\code{\link{set_par_list}}}
}
\value{
data.frame
}
\description{
Caution! This function will take much time, because it is only
optimised for small sequences.
}
\details{
This function is based on samtools to get the reference to the
hit. Further the consensus is generated by adding blanks "-" to
the start and the end of the reads to cover the full
reference. Therefore, it will be very slow, if the reference
becomes very large.
}
\author{
Jochen Kruppa
}
