% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pauda_hits.R
\name{get_pauda_hits}
\alias{get_pauda_hits}
\title{Test}
\usage{
get_pauda_hits(file, par_list)
}
\arguments{
\item{file}{}

\item{par_list}{}
}
\value{
data.frame
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
