% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_sub_files}
\alias{get_sub_files}
\title{TEst}
\usage{
get_sub_files(sample, batch = NULL, in_dir = NULL, out_dir = NULL,
  name = NULL, gz = FALSE)
}
\arguments{
\item{sample}{}

\item{batch}{}

\item{in_dir}{}

\item{out_dir}{}

\item{name}{}

\item{gz}{}
}
\value{
file.list
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
