% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is.named}
\alias{is.named}
\title{Small tester if a object is named}
\usage{
is.named(file)
}
\arguments{
\item{file}{A file string}
}
\value{
logical
}
\description{
Small tester if a object or file is named
}
\details{
Small tester if a object or file is named
}
\author{
Jochen Kruppa
}
