% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dna_ref.R
\name{map_dna_ref}
\alias{map_dna_ref}
\title{Test}
\usage{
map_dna_ref(infile, outfile, par_list, min_hit = 5, all = FALSE)
}
\arguments{
\item{infile}{}

\item{outfile}{}

\item{par_list}{}

\item{min_hit}{}

\item{all}{}
}
\value{
data.frame
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
