% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_dna_plot.R
\name{mapping_dna_plot}
\alias{mapping_dna_plot}
\title{Mapping the viral read hits to the reference}
\usage{
mapping_dna_plot(genebank_id, consensus_df = NULL, dna_alignment_df,
  pep_alignment_df, aa_info_df, viral_info_df, sample_info = NULL,
  pdf_file = NULL, par_list)
}
\arguments{
\item{genebank_id}{Genebank IDs of the hits}

\item{consensus_df}{Consensus data frame generated by
\code{\link{get_consensus_df}}}

\item{dna_alignment_df}{DNA alignment data frame generated by
\code{\link{map_dna_ref}}}

\item{pep_alignment_df}{AA alignment data frame generated by
\code{\link{map_pep_ref}}}

\item{aa_info_df}{SQLlite database from the
\code{\link{setup_aa_info_sample_sqlite}}}

\item{viral_info_df}{SQLlite database from the
\code{\link{setup_species_info_sqlite}}}

\item{sample_info}{Sample info data frame generated by
\code{\link{build_sample_info}}}

\item{pdf_file}{The output file, where the plot should be saved}

\item{par_list}{Parameter given by the par_list(), see
\code{\link{set_par_list}}}
}
\description{
A plotting function for the mapping results of the found viral strains
}
\details{
This is a realtivly complex function. First the genebank hit ids
must be provided. All IDs, which are provided, must be found in
the other options. Second a potential consensus data frame, where
the consensus information of the reads to the viral reference is
given, see \code{\link{get_consensus_df}} for more
information. Then, the alignment of the DNA and AA mapping must be
provided, see \code{\link{map_dna_ref}} and
\code{\link{map_pep_ref}} for more information. Further, the
SQLlite database for the viral info and the amino acid info must
be given.
}
\author{
Jochen Kruppa
}
