% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NHS_10001_consensus}
\alias{NHS_10001_consensus}
\title{Output of get_consensus_df}
\format{A list with two entries:

$consensus_df
\describe{
  \item{genebank_id}{Genbank ID of the reference viral strain}
  \item{con_prob}{Coverage probability. How unique is the base?}
  \item{pos}{Position on the reference}
  \item{bool}{Mathc between the reference and the consensus sequence}
}

$entropy_df
\describe{
  \item{genebank_id}{Genbank ID of the reference viral strain}
  \item{entropy}{Entropy of the consensus reference}
}

$run_len_df
\describe{
  \item{genebank_id}{Genbank ID of the reference viral strain}
  \item{min}{Minimum run of bases to the reference}
  \item{first}{First quartile of the runs to the reference}
  \item{median}{Median of the runs to the reference}
  \item{mean}{Mean of the runs to the reference}
  \item{third}{Third quartile of the runs to the reference}
  \item{max}{Maximum run of bases to the reference}
  \item{run_to_ref}{Maximal run to the reference}
  \item{run_not_to_ref}{Maximal run _not_ to the reference}
}

$median_con_prob_df
\describe{
  \item{genebank_id}{Genbank ID of the reference viral strain}
  \item{min}{Minimum of the consensus_df$con_prob}
  \item{first}{First quartile of the consensus_df$con_prob}
  \item{median}{Median of the consensus_df$con_prob}
  \item{mean}{Mean of the consensus_df$con_prob}
  \item{third}{Third quartile of the consensus_df$con_prob}
  \item{max}{Maximum of the consensus_df$con_prob}
}

$coverage_equal_ref_df
\describe{
  \item{genebank_id}{Genbank ID of the reference viral strain}
  \item{coverage_equal_ref}{How many of the consensus base match the reference base?}
}

$coverage_overall_ref_df
\describe{
  \item{genebank_id}{Genbank ID of the reference viral strain}
  \item{coverage_overall_ref}{Ho much of the reference is covered by the consensus sequence}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/genbank}
}
\usage{
NHS_10001_consensus
}
\description{
A example output of the \code{\link{get_consensus_df}} function.
}
\keyword{datasets}
