% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NHS_10001_count_tbl}
\alias{NHS_10001_count_tbl}
\title{Dataset}
\format{A data.frame with five entries:
\describe{
  \item{genebank_id}{Genbank ID}
  \item{dna_reads}{Count of DNA reads mapped to the reference Genbank ID}
  \item{dna_rank}{Rank of the DNA read count}
  \item{amino_reads}{Count of AA reads mapped to the reference Genbank ID}
  \item{amino_rank}{Rank of the AA read count}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/genbank}
}
\usage{
NHS_10001_count_tbl
}
\description{
A example output of \code{\link{ord_findings}}.
}
\keyword{datasets}
