% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NHS_10001_map_aa_list}
\alias{NHS_10001_map_aa_list}
\title{Dataset}
\format{A list with two entries:
\describe{
  \item{R1}{Fastq reads object with 10000 paired reads of R1}
  \item{R2}{Fastq reads object with 10000 paired reads of R2}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/genbank}
}
\usage{
NHS_10001_map_aa_list
}
\description{
Test
}
\keyword{datasets}
