% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NHS_10001_map_dna_list}
\alias{NHS_10001_map_dna_list}
\title{Dataset}
\format{A list with two data.frame:
$count_df
\describe{
  \item{genebank_id}{Genbank reference ID}
  \item{count_read}{Number of reads mapped to the genebank ID reference}
}
$alignment_df
\describe{
  \item{genebank_id}{Genbank reference ID}
  \item{qname}{Read name i.e. query name}
  \item{strand}{Strand}
  \item{read_length}{Read length}
  \item{pos_start}{Poistion start on the reference DNA sequence}
  \item{pos_end}{Poistion end on the reference DNA sequence}
  \item{mapq}{Mapping quality}
  \item{seq_id}{Additional identifier}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/genbank}
}
\usage{
NHS_10001_map_dna_list
}
\description{
A example output of \code{\link{map_dna_ref}}.
}
\keyword{datasets}
