% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ord_findings}
\alias{ord_findings}
\title{Test}
\usage{
ord_findings(map_dna_list, map_pep_list)
}
\arguments{
\item{map_dna_list}{}

\item{map_pep_list}{}
}
\value{
data.frame
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
