% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot_time_sample}
\alias{plot_time_sample}
\title{Test}
\usage{
plot_time_sample(batch_fastq_files)
}
\arguments{
\item{batch_fastq_files}{}
}
\value{
ggplot object
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
