% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna_seq.R
\name{random_dna_seq}
\alias{random_dna_seq}
\title{title}
\usage{
random_dna_seq(inFile, outFile, k, n = 1, clean = TRUE, tmpDir,
  fasta_formatter = NULL, fasta_uShuffle = NULL)
}
\arguments{
\item{inFile}{}

\item{outFile}{}

\item{k}{}

\item{n}{}

\item{clean}{}

\item{tmpDir}{}

\item{fasta_formatter}{}

\item{fasta_uShuffle}{}
}
\value{
Null
}
\description{
descr
}
\details{
details
}
\author{
Jochen Kruppa
}
