% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_batch_job.R
\name{run_batch_job}
\alias{run_batch_job}
\title{Test}
\usage{
run_batch_job(batch_in_dir, batch_out_dir, par_list, run_samples = NULL)
}
\arguments{
\item{batch_in_dir}{}

\item{batch_out_dir}{}

\item{par_list}{}

\item{run_samples}{}
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
