% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_functions.R
\name{set_par_list}
\alias{set_par_list}
\title{TEst}
\usage{
set_par_list(index_genome_dir, index_amino_dir, qc = TRUE,
  set_mapper = "bowtie", ref_seq_file = "", prot_info_df = "",
  species_info_df = "")
}
\arguments{
\item{index_genome_dir}{}

\item{index_amino_dir}{}

\item{qc}{}

\item{set_mapper}{}

\item{ref_seq_file}{}

\item{prot_info_df}{}

\item{species_info_df}{}
}
\value{
foo
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
