% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_functions.R
\name{set_program_list}
\alias{set_program_list}
\title{Test}
\usage{
set_program_list(bowtie_dir, pauda_dir, samtools_dir, trimmomatic_dir)
}
\arguments{
\item{bowtie_dir}{}

\item{pauda_dir}{}

\item{samtools_dir}{}

\item{trimmomatic_dir}{}
}
\value{
program_list
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
