% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{setup_aa_info_sqlite}
\alias{setup_aa_info_sqlite}
\title{title}
\usage{
setup_aa_info_sqlite(ncbi_aa_seq, db_file)
}
\arguments{
\item{ncbi_aa_seq}{}

\item{db_file}{}
}
\value{
Null
}
\description{
desc
}
\details{
detail
}
\author{
Jochen Kruppa
}
