% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{setup_species_info_sqlite}
\alias{setup_species_info_sqlite}
\title{Function to build up the species_info_sqlite database}
\usage{
setup_species_info_sqlite(genebank_id, length, description, tax_id = NA,
  mol_type = NA, strain = NA, accession = NA, organism = NA, db_file)
}
\arguments{
\item{genebank_id}{Vector of genebank_ids}

\item{length}{Length vector of the sequence}

\item{description}{Description vector of the genebank_ids}

\item{tax_id}{NCBI tax_id connected with the genebank_ids}

\item{mol_type}{Molecular type}

\item{strain}{Strain}

\item{accession}{Accession number}

\item{organism}{Organsim}

\item{db_file}{File path to save the sqlite3 database}
}
\description{
The function generates the SQlite databse needed for the plotting.
}
\details{
The function generates the SQlite databse needed for the plotting
of the hit seqeunces.
}
\author{
Jochen Kruppa
}
