% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strain_info_db}
\alias{strain_info_db}
\title{Dataset}
\format{A dataset containing 287 viral strains and information:
\describe{
  \item{1st}{Genbank ID}
  \item{2nd}{Molecular type} 
  \item{3rd}{Strain: RNA, DNA or other}
  \item{4th}{Taxonomic ID}
  \item{5th}{Additional accession number}
  \item{6th}{Length}
  \item{7th}{Organism}
  \item{8th}{Description}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/genbank}
}
\usage{
strain_info_db
}
\description{
A dataset containing 287 viral strains and information
}
\keyword{datasets}
