% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{talk}
\alias{talk}
\title{Advanced message functionality by time stamp}
\usage{
talk(...)
}
\arguments{
\item{...}{Character string}
}
\value{
Message with time stamp
}
\description{
Advanced message() functionality by time stamp
}
\details{
Advanced message() functionality by time stamp using the function
\code{\link{TIME}}.
}
\author{
Jochen Kruppa
}
