% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tbl_sqlite_check}
\alias{tbl_sqlite_check}
\title{Test function to check the sqlite2 databases for consistency}
\usage{
tbl_sqlite_check(object)
}
\arguments{
\item{object}{par_list object}
}
\value{
boolean
}
\description{
A generic function to check the par_list entries
}
\details{
A generic function to check the par_list entries
}
\author{
Jochen Kruppa
}
