% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{TIME}
\alias{TIME}
\title{Small wrapper to extract the time from Sys.time()}
\usage{
TIME(...)
}
\arguments{
\item{...}{No parameters are needed}
}
\value{
System time in %b %d %X
}
\description{
Small wrapper to extract the time from Sys.time()
}
\details{
Small wrapper to extract the time from Sys.time()
}
\author{
Jochen Kruppa
}
