% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artificial_genome_mapping.R
\name{artificial_genome_mapping}
\alias{artificial_genome_mapping}
\title{Test}
\usage{
artificial_genome_mapping(in_file, out_file, par_list)
}
\arguments{
\item{in_file}{}

\item{out_file}{}

\item{par_list}{}
}
\value{
Test
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
