% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_host}
\alias{check_host}
\title{Test}
\usage{
check_host(file, force = FALSE)
}
\arguments{
\item{file}{}

\item{force}{}
}
\value{
data.frame
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
