% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_filter.R
\name{coverage_filter}
\alias{coverage_filter}
\title{Test}
\usage{
coverage_filter(hit_id, map_dna_list, par_list, out_file)
}
\arguments{
\item{hit_id}{}

\item{map_dna_list}{}

\item{par_list}{}

\item{out_file}{}
}
\value{
tibble
}
\description{
Desc
}
\details{
Detail
}
\author{
Jochen Kruppa
}
