% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_obj.R
\name{empty_map_pep_list}
\alias{empty_map_pep_list}
\title{Test}
\usage{
empty_map_pep_list(...)
}
\arguments{
\item{...}{}
}
\value{
data.frame
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
