% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ete3.R
\name{ete2_plot}
\alias{ete2_plot}
\title{Test}
\usage{
ete2_plot(tax_ids, out_file, pdf_file, clean = FALSE)
}
\arguments{
\item{tax_ids}{}

\item{out_file}{}

\item{pdf_file}{}

\item{clean}{}
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
