% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{fastq_quality_control}
\alias{fastq_quality_control}
\title{Main fastq quality function}
\usage{
fastq_quality_control(inFile, leading = 10, trailing = 10, minlength = 50,
  illumninaclip, par_list)
}
\arguments{
\item{inFile}{Fastq infile}

\item{leading}{Numeric. Remove leading low quality or N bases
(below quality 10 [default])}

\item{trailing}{Numeric. Remove trailing low quality or N bases
(below quality 10 [default])}

\item{minlength}{Min read length to keep. 50 [default]}

\item{illumninaclip}{Remove the adapter from illumina by TruSeq2
or TruSeq3 protocol}

\item{par_list}{}

\item{log_file}{write the trimmomatic output to file}
}
\description{
Main fastq quality function
}
\details{
Main fastq quality function. Calls the function
fastq_quality_filter and fastq_trimmer. Stores files in a tempDir.
}
\author{
Jochen Kruppa
}
