% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_decoy_reads.R
\name{generate_decoy_reads}
\alias{generate_decoy_reads}
\title{test}
\usage{
generate_decoy_reads(decoy_seq, paired = TRUE, read_num = 10000,
  read_length = 50, read_gap = 50, split = 10)
}
\arguments{
\item{decoy_seq}{}

\item{paired}{}

\item{read_num}{}

\item{read_length}{}

\item{read_gap}{}

\item{split}{}
}
\description{
est desc
}
\details{
test detail
}
\author{
Jochen Kruppa
}
