% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_decoy_reads.R
\name{generate_read}
\alias{generate_read}
\title{gb}
\usage{
generate_read(decoy_seq, files, paired, read_length, read_gap)
}
\arguments{
\item{decoy_seq}{}

\item{files}{}

\item{paired}{}

\item{read_length}{}

\item{read_gap}{}
}
\description{
Test desc
}
\details{
Test detail
}
\author{
Jochen Kruppa
}
