% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bowtie2_cmd.R
\name{get_bowtie2_cmd}
\alias{get_bowtie2_cmd}
\title{Test}
\usage{
get_bowtie2_cmd(inFile, samOutFile, referenceDir, method, p = 32,
  all = FALSE)
}
\arguments{
\item{inFile}{}

\item{samOutFile}{}

\item{referenceDir}{}

\item{method}{}

\item{p}{}

\item{all}{}
}
\value{
character
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
