% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{get_consensus_df}
\alias{get_consensus_df}
\title{Test}
\usage{
get_consensus_df(hits, out_file, par_list)
}
\arguments{
\item{hits}{}

\item{out_file}{}

\item{par_list}{}
}
\value{
data.frame
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
