% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_seq_chunks_by_threshold}
\alias{get_seq_chunks_by_threshold}
\title{test}
\usage{
get_seq_chunks_by_threshold(seqs, threshold)
}
\arguments{
\item{seqs}{}

\item{threshold}{}
}
\value{
null
}
\description{
desc
}
\details{
detail
}
\author{
Jochen Kruppa
}
