% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_pep_ref.R
\name{map_pep_ref}
\alias{map_pep_ref}
\title{Test}
\usage{
map_pep_ref(infile, outfile, par_list)
}
\arguments{
\item{infile}{}

\item{outfile}{}

\item{par_list}{}
}
\value{
data.frame
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
