% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_dna_plot.R
\name{mapping_dna_plot}
\alias{mapping_dna_plot}
\title{Test}
\usage{
mapping_dna_plot(genebank_id, consensus_df = NULL, dna_alignment_df,
  pep_alignment_df, aa_info_df, viral_info_df, sample_info = NULL,
  pdf_file = NULL, par_list)
}
\arguments{
\item{genebank_id}{}

\item{consensus_df}{}

\item{dna_alignment_df}{}

\item{pep_alignment_df}{}

\item{aa_info_df}{}

\item{viral_info_df}{}

\item{sample_info}{}

\item{pdf_file}{}

\item{par_list}{}
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
