% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ord_findings}
\alias{ord_findings}
\title{Order the hits of the mapping}
\usage{
ord_findings(map_dna_list, map_pep_list)
}
\arguments{
\item{map_dna_list}{List with the DNA read hits from \code{\link{map_dna_ref}}}

\item{map_pep_list}{List with the AA read hits from \code{\link{map_pep_ref}}}
}
\value{
Ordered data.frame with the counts
}
\description{
The hits from the DNA and AA mapping are merged and ordered by their combined rank.
}
\details{
The hits from the DNA and AA mapping are merged and ordered by their combined rank.
}
\author{
Jochen Kruppa
}
