% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot_time_sample}
\alias{plot_time_sample}
\title{Plot the processed files in a dir}
\usage{
plot_time_sample(batch_fastq_files)
}
\arguments{
\item{batch_fastq_files}{A list of files to be parsed}
}
\value{
ggplot object
}
\description{
This an experimental function, which is very specific
}
\details{
In the daily workflow, some descriptions must be generated. This
is one of the functions.
}
\author{
Jochen Kruppa
}
