% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam2bam.R
\name{sam2bam_serial}
\alias{sam2bam_serial}
\title{Test}
\usage{
sam2bam_serial(sam_files, keep = FALSE, index_sort = FALSE)
}
\arguments{
\item{sam_files}{}

\item{keep}{}

\item{index_sort}{}
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
