% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_functions.R
\name{set_par_list}
\alias{set_par_list}
\title{S4 class of the parameter list}
\usage{
set_par_list(index_genome_dir, index_amino_dir, qc = TRUE,
  set_mapper = "bowtie", ref_seq_file = "", prot_info_df = "",
  species_info_df = "")
}
\arguments{
\item{index_genome_dir}{path to the index of the reference genome
generated by Bowtie2 or Star}

\item{index_amino_dir}{path to the index of the amino acid
reference genome generated by Pauda}

\item{qc}{should the quality trimming used [true]}

\item{set_mapper}{which mapper should be used [bowtie]}

\item{ref_seq_file}{path to the reference sequence file in fasta
format}

\item{prot_info_df}{sqlite3 database of the protein information,
see \code{\link{setup_aa_info_sample_sqlite}} for setup}

\item{species_info_df}{sqlite3 database of the species/strain
information, see \code{\link{setup_species_info_sqlite}} for
setup}
}
\value{
A S4 class with the following slots [default values]
\describe{
\item{tax = "logical"}{should the tax trees are generated [true]}
\item{qc = "logical"}{should the quality control by trimmomatic be run [true]}
\item{clean = "logical"}{should intermediate files be deleted [true]} 
\item{paired = "logical"}{should paired read infiles assumed [false], but will be checked internally}
\item{plot = "logical"}{should the final visualization of the mapped reads plotted [true]}
\item{black_list = "character"}{list of corrupt NCBI GenBAnk IDs [Null]}
\item{gen_prot_sep = "character"}{the seperator, how the genebank_id is seperated prom the prot_id [_]}
\item{check_host = "logical"}{should the host be checked by blastn [true]}
\item{run_dna_mapping = "logical"}{should the DNA mapping be run [true]}
\item{run_pep_mapping = "logical"}{should the amino acid mapping be run? [true]}
\item{consensus = "logical"}{should the consensus of the reads be generated [true]}
\item{tmp_dir = "character"}{path to the temp dir [/home/temp]}
\item{sql_dir = "character"}{path to the sql dir [/home/sql]}
\item{index_genome_dir = "character"}{path to the genome index generated by Star or Bowtie2}
\item{index_amino_dir = "character"}{path to the amino index generated by Pauda}
\item{ref_seq_file = "character"}{path to the reference.fasta}
\item{prot_info = "tbl_dbi"}{sqlite3 database of the protein information}
\item{species_info = "tbl_dbi"}{sqlite3 database of the species/strain information}
\item{plot_id = "character"}{should only a subset of genebank_ids be plotted?}
\item{mapper = "character"}{sould the Bowtie2 or Star mapper be used [bowtie]}
\item{num_plot = "numeric"}{how many plots of the TOP`num_plot` should be generated [25]}
\item{ncore = "numeric"}{how many cores should be used}
\item{pdf_file = "character"}{name of the final plot pdf file}
}
}
\description{
Generates the S4 class of the parameter list
}
\details{
The S4 class has some validity functions to check the input
parameters. The access of the functions always runs by
par_list["value"]. If no S4 class is wanted, the user could also
build a own list object with the given names.
}
\author{
Jochen Kruppa
}
