% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{setup_aa_info_sample_sqlite}
\alias{setup_aa_info_sample_sqlite}
\title{Test}
\usage{
setup_aa_info_sample_sqlite(prot_id, genebank_id, prot_start, prot_end, db_file)
}
\arguments{
\item{prot_id}{}

\item{genebank_id}{}

\item{prot_start}{}

\item{prot_end}{}

\item{db_file}{}
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
