% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{setup_aa_info_sqlite}
\alias{setup_aa_info_sqlite}
\title{Function to build up the aa_info_sqlite database}
\usage{
setup_aa_info_sqlite(ncbi_aa_seq, db_file)
}
\arguments{
\item{ncbi_aa_seq}{A AAStringSet containing amino acids}

\item{db_file}{File path to the sqlite3 database}
}
\value{
Null
}
\description{
The function generates the SQlite databse needed for the plotting.
}
\details{
The function generates the SQlite databse needed for the plotting
of the hit sequences.
}
\author{
Jochen Kruppa
}
