% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{setup_species_info_sqlite}
\alias{setup_species_info_sqlite}
\title{Test}
\usage{
setup_species_info_sqlite(genebank_id, length, description, tax_id = NA,
  mol_type = NA, strain = NA, accession = NA, organism = NA, db_file)
}
\arguments{
\item{genebank_id}{}

\item{length}{}

\item{description}{}

\item{tax_id}{}

\item{mol_type}{}

\item{strain}{}

\item{accession}{}

\item{organism}{}

\item{db_file}{}
}
\description{
Test
}
\details{
Test
}
\author{
Jochen Kruppa
}
