% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strain_info_db}
\alias{strain_info_db}
\title{Dataset}
\format{A fasta file with 1000 entries. The names are split by ' ' meaning:
\describe{
  \item{1st}{Genbank ID}
  \item{2nd}{Genbank ID}
  \item{3rd - End}{Description}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/genbank}
}
\usage{
strain_info_db
}
\description{
A dataset containing 1000 DNA sequences from the NCBI GenBank.
}
\keyword{datasets}
