% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{talk}
\alias{talk}
\title{Message function with time}
\usage{
talk(...)
}
\arguments{
\item{...}{}
}
\value{
Message with time
}
\description{
Message function with time
}
\details{
Message function with time
}
\author{
Jochen Kruppa
}
